requirements.txt
Flask==3.1.2
Jinja2==3.1.6
Werkzeug==3.1.3
itsdangerous==2.2.0
MarkupSafe==3.0.3
blinker==1.9.0
click==8.3.0
openpyxl==3.1.5
et_xmlfile==2.0.0
colorama==0.4.6
bcrypt==4.2.1
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
apscheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.0

